# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type EmployeeRole = typeof EmployeeRole[keyof typeof EmployeeRole];


export const EmployeeRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  email: string;
  role: EmployeeRole;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  employee: Employee;
}

export type EmployeeInputRole = typeof EmployeeInputRole[keyof typeof EmployeeInputRole];


export const EmployeeInputRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface EmployeeInput {
  employeeId: string;
  name: string;
  email: string;
  password: string;
  role: EmployeeInputRole;
}

export type EmployeeUpdateRole = typeof EmployeeUpdateRole[keyof typeof EmployeeUpdateRole];


export const EmployeeUpdateRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface EmployeeUpdate {
  name?: string;
  email?: string;
  role?: EmployeeUpdateRole;
  password?: string;
}

export interface QrToken {
  id: number;
  token: string;
  expiresAt: string;
  createdAt: string;
}

export interface ClockInInput {
  token: string;
  latitude: number;
  longitude: number;
  /** Base64-encoded selfie image (data URL) */
  selfieBase64?: string;
}

export interface ClockOutInput {
  token: string;
  latitude: number;
  longitude: number;
  /** Base64-encoded selfie image (data URL) */
  selfieBase64?: string;
}

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  date: string;
  clockIn: string;
  /** @nullable */
  clockOut?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  selfieUrl?: string | null;
  createdAt: string;
}

export interface AttendanceOrNull {
  record: AttendanceRecord | null;
}

export interface AttendanceWithEmployee {
  id: number;
  employeeId: number;
  date: string;
  clockIn: string;
  /** @nullable */
  clockOut?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  selfieUrl?: string | null;
  createdAt: string;
  employee: Employee;
}

export interface DashboardStats {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  lateToday: number;
  avgHoursThisMonth: number;
  recentAttendance: AttendanceWithEmployee[];
}

export type ListEmployeesParams = {
search?: string;
};

export type GetMyAttendanceHistoryParams = {
/**
 * Month in YYYY-MM format
 */
month?: string;
};

export type ListAttendanceParams = {
/**
 * Filter by date YYYY-MM-DD
 */
date?: string;
employeeId?: number;
/**
 * Filter by month YYYY-MM
 */
month?: string;
};

export type ExportAttendanceParams = {
/**
 * Month in YYYY-MM format
 */
month?: string;
employeeId?: number;
};