# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Employee Attendance System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for employees and admins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance/clock-in:
    post:
      operationId: clockIn
      tags: [attendance]
      summary: Clock in using QR token (with GPS + selfie)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInInput"
      responses:
        "201":
          description: Clocked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Invalid token, already clocked in, or out of radius
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/clock-out:
    post:
      operationId: clockOut
      tags: [attendance]
      summary: Clock out using QR token (with GPS + selfie)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutInput"
      responses:
        "200":
          description: Clocked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Invalid token or not clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [attendance]
      summary: Get today's attendance for the current employee
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today's attendance record (null if not clocked in)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceOrNull"
  /attendance/my-history:
    get:
      operationId: getMyAttendanceHistory
      tags: [attendance]
      summary: Get attendance history for the current employee
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "Month in YYYY-MM format"
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List all attendance records (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            description: "Filter by date YYYY-MM-DD"
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
            description: "Filter by month YYYY-MM"
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceWithEmployee"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /admin/attendance/export:
    get:
      operationId: exportAttendance
      tags: [admin]
      summary: Export attendance records as CSV
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "Month in YYYY-MM format"
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: CSV data
          content:
            text/csv:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResult:
      type: object
      required: [token, employee]
      properties:
        token:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
    Employee:
      type: object
      required: [id, employeeId, name, email, role, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, employee]
        createdAt:
          type: string
          format: date-time
...
    ClockInInput:
      type: object
      required: [token, latitude, longitude]
      properties:
        token:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        selfieBase64:
          type: string
          description: Base64-encoded selfie image (data URL)
    ClockOutInput:
      type: object
      required: [token, latitude, longitude]
      properties:
        token:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        selfieBase64:
          type: string
          description: Base64-encoded selfie image (data URL)
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, clockIn, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        date:
          type: string
          format: date
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: ["string", "null"]
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        selfieUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AttendanceOrNull:
      type: object
      required: [record]
      properties:
        record:
          oneOf:
            - $ref: "#/components/schemas/AttendanceRecord"
            - type: "null"
    AttendanceWithEmployee:
      type: object
      required: [id, employeeId, date, clockIn, createdAt, employee]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        date:
          type: string
          format: date
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: ["string", "null"]
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        selfieUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        employee:
          $ref: "#/components/schemas/Employee"
    DashboardStats:
      type: object
      required: [totalEmployees, presentToday, absentToday, lateToday, avgHoursThisMonth, recentAttendance]
      properties:
        totalEmployees:
          type: integer
        presentToday:
          type: integer
        absentToday:
          type: integer
        lateToday:
          type: integer
        avgHoursThisMonth:
          type: number
        recentAttendance:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceWithEmployee"