# lib/api-zod/src/generated/types/attendanceOrNull.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecord } from './attendanceRecord';

export interface AttendanceOrNull {
  record: AttendanceRecord | null;
}