# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from './employee';

export interface AuthResult {
  token: string;
  employee: Employee;
}