# lib/api-zod/src/generated/types/clockOutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface ClockOutInput {
  token: string;
  latitude: number;
  longitude: number;
  /** Base64-encoded selfie image (data URL) */
  selfieBase64?: string;
}