# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeRole } from './employeeRole';

export interface Employee {
  id: number;
  employeeId: string;
  name: string;
  email: string;
  role: EmployeeRole;
  createdAt: Date;
}